Added clearExistingFile(bool) input to TripCoordinateReadWrite Constructor.

+++ FrameWorkApp/FrameWorkApp/Helper Classes/TripCoordinateReadWrite.cs
@@ -11,11 +11,13 @@ namespace FrameWorkApp
 		String libraryCache;
 		String filePath;
 
-		public TripCoordinateReadWrite ()
+		public TripCoordinateReadWrite (Boolean clearExistingFile)
 		{
 			libraryCache= Path.Combine (Environment.GetFolderPath (Environment.SpecialFolder.MyDocuments), "..", "Library", "Caches");
 			filePath=Path.Combine (libraryCache, "currentTrip_SafeDrivingMate.txt");
-			File.WriteAllText(filePath, "");
+			if (clearExistingFile) {
+				File.WriteAllText (filePath, "");
+			}
 		}
 
 
@@ -24,7 +26,6 @@ namespace FrameWorkApp
 			FileStream currentTripFile_FileStream = File.Open (filePath,FileMode.Append);
 			StreamWriter currentTripFile_SteamWriter = new StreamWriter (currentTripFile_FileStream);
 
-			Console.WriteLine (filePath);
 			currentTripFile_SteamWriter.WriteLine (newCoordiante.Latitude+","+newCoordiante.Longitude);
 
 			currentTripFile_SteamWriter.Close ();
